Cherry picking: Temporarily increase DTLS buffer size to 2.
It's not expected this will make a difference, since the packet should
be read from the queue if possible as soon as it's added to it.

But we're doing this as an added precaution in case we overlooked
something. See linked bug.

Bug: chromium:1063834
Change-Id: I7a3a6d86a97683cbcbeed5ef1aaa8090cf6bf8c0
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/172661
Commit-Queue: Taylor <deadbeef@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Original-Commit-Position: refs/heads/master@{#30990}
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/172961
Cr-Commit-Position: refs/branch-heads/4103@{#1}
Cr-Branched-From: 04c1b445019e10e54b96f70403d25cc54215faf3-refs/heads/master@{#30987}